Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/same-origin/attr-referrer/same-origin/http-http/iframe-tag/swap-origin-redirect/same-origin-insecure.http-expected.txt



FAIL The referrer URL is omitted when a
                                 document served over http requires an http
                                 sub-resource via iframe-tag using the attr-referrer
                                 delivery method with swap-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'omitted'. expected (undefined) undefined but got (string) "http://localhost:8800/referrer-policy/same-origin/attr-referrer/same-origin/http-http/iframe-tag/swap-origin-redirect/same-origin-insecure.http.html"
